# trading-chats-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "凌期AI辅助期货挑选品种和持仓止盈止损分析程序API文档",
        "title": "凌期AI辅助期货挑选品种和持仓止盈止损分析程序",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/ai-responses": {
            "get": {
                "description": "获取所有AI响应信息列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI响应信息"
                ],
                "summary": "获取所有AI响应信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ai-responses/batch": {
            "get": {
                "description": "根据批次ID获取AI响应信息列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI响应信息"
                ],
                "summary": "根据批次ID获取AI响应信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "批次ID",
                        "name": "batch_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ai-responses/generate": {
            "post": {
                "description": "根据提示词模版和参数生成批次AI响应",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI响应信息"
                ],
                "summary": "生成批次AI响应",
                "parameters": [
                    {
                        "description": "生成批次AI响应请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GenerateAIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ai-responses/latest": {
            "get": {
                "description": "获取最近一次状态为 completed 的完整批次 AI 响应列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI响应信息"
                ],
                "summary": "获取最近一次成功的批次数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ai-responses/{id}": {
            "get": {
                "description": "根据ID获取AI响应信息详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI响应信息"
                ],
                "summary": "根据ID获取AI响应信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "AI响应信息ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.GenerateAIRequest": {
            "type": "object",
            "properties": {
                "param1": {
                    "type": "string",
                    "example": "https://api.example.com/data1"
                },
                "param2": {
                    "type": "string",
                    "example": "https://api.example.com/data2"
                },
                "template_id": {
                    "type": "string",
                    "example": "69b4e08fbd4f24a37ff0a0cc"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}